Concluido, Permisão user funcionando

+++ usuario/src/main/java/com/shadow/desafio/controllers/UsuariosController.java
@@ -1,11 +1,9 @@
 package com.shadow.desafio.controllers;
 
 import com.shadow.desafio.dtos.LoginDto;
-import com.shadow.desafio.dtos.UsuariosDto;
 import com.shadow.desafio.entities.Usuarios;
 import com.shadow.desafio.service.UsuariosService;
 import com.shadow.desafio.service.exceptions.MessageExceptionHandler;
-import org.springframework.beans.BeanUtils;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,7 +11,6 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.Date;
-import java.util.Optional;
 import java.util.UUID;
 
 @RestController
@@ -32,9 +29,9 @@ public class UsuariosController {
         if (!usuariosService.validarToken(token)) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new MessageExceptionHandler(new Date(), HttpStatus.UNAUTHORIZED.value(), "Token Inválido" ));
         }
-//        if (!usuariosService.getTypeUser(token).equals("ADMIN")){
-//            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new MessageExceptionHandler(new Date(), HttpStatus.FORBIDDEN.value(), "Usuário sem permisão" ));
-//        }
+        if (!usuariosService.getTypeUser(token).equals("ADMIN")){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new MessageExceptionHandler(new Date(), HttpStatus.FORBIDDEN.value(), "Usuário sem permisão" ));
+        }
 
         if (usuariosService.existsByEmail(usuarios.getEmail())){
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new MessageExceptionHandler(new Date(), HttpStatus.NOT_FOUND.value(),"E-mail já cadastrado" ));
@@ -55,39 +52,59 @@ public class UsuariosController {
         if (!usuariosService.validarToken(token)) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new MessageExceptionHandler(new Date(), HttpStatus.UNAUTHORIZED.value(), "Token Inválido" ));
         }
-//        if (!usuariosService.getTypeUser(token).equals("ADMIN")){
-//            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new MessageExceptionHandler(new Date(), HttpStatus.FORBIDDEN.value(), "Usuário sem permisão" ));
-//        }
-
+        if (!usuariosService.getTypeUser(token).equals("ADMIN")){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new MessageExceptionHandler(new Date(), HttpStatus.FORBIDDEN.value(), "Usuário sem permisão" ));
+        }
 
         return ResponseEntity.status(HttpStatus.OK).body(usuariosService.findAll());
     }
     @GetMapping(value = "/{codigoID}") // Buscar por Id.
     public ResponseEntity<Object> buscarID(@PathVariable UUID codigoID,
                                             @RequestHeader(HttpHeaders.AUTHORIZATION)String token) {
+        if (!usuariosService.validarToken(token)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new MessageExceptionHandler(new Date(), HttpStatus.UNAUTHORIZED.value(), "Token Inválido" ));
+        }
+        if (!usuariosService.getTypeUser(token).equals("ADMIN")){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new MessageExceptionHandler(new Date(), HttpStatus.FORBIDDEN.value(), "Usuário sem permisão" ));
+        }
          return new ResponseEntity<>(usuariosService.findById(codigoID), HttpStatus.OK);
     }
     @DeleteMapping("/{codigoID}") // Delete por ID.
     public ResponseEntity<Object> delete(@PathVariable(value = "codigoID") UUID codigoID,
                                          @RequestHeader(HttpHeaders.AUTHORIZATION)String token){
+        if (!usuariosService.validarToken(token)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new MessageExceptionHandler(new Date(), HttpStatus.UNAUTHORIZED.value(), "Token Inválido" ));
+        }
+        if (!usuariosService.getTypeUser(token).equals("ADMIN")){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new MessageExceptionHandler(new Date(), HttpStatus.FORBIDDEN.value(), "Usuário sem permisão" ));
+        }
+
         Usuarios usuarios = usuariosService.findById(codigoID);
         usuariosService.delete(usuarios);
         return ResponseEntity.status(HttpStatus.OK).body(new MessageExceptionHandler(new Date(),HttpStatus.OK.value(),"Usuário deletado"));
     }
     @PutMapping("/{id}")
     public ResponseEntity<Object> atualizar(@PathVariable(value = "id") UUID codigoID,
-                                            @RequestBody @Valid UsuariosDto usuariosDto,
+                                            @RequestBody @Valid Usuarios usuarios,
                                             @RequestHeader(HttpHeaders.AUTHORIZATION)String token){
-        Optional<Usuarios> usuariosOptional = Optional.ofNullable(usuariosService.findById(codigoID));
-        if (!usuariosOptional.isPresent()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário não encontrado.");
+
+        if (!usuariosService.validarToken(token)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new MessageExceptionHandler(new Date(), HttpStatus.UNAUTHORIZED.value(), "Token Inválido" ));
         }
-        var usuarios = new Usuarios();
-        BeanUtils.copyProperties(usuariosDto, usuarios); /* conversão de Dto para entity*/
-        usuarios.setCodigoID(usuariosOptional.get().getCodigoID()); /* Setando Id para permanecer o mesmo */
+        if (!usuariosService.getTypeUser(token).equals("ADMIN")){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new MessageExceptionHandler(new Date(), HttpStatus.FORBIDDEN.value(), "Usuário sem permisão" ));
+        }
+        Usuarios user = usuariosService.findById(codigoID);
+        user.setNome(user.getNome());
+        user.setCpf(user.getCpf());
+        user.setEmail(user.getEmail());
+        user.setSenha(user.getSenha());
+        user.setTipoUsuario(user.getTipoUsuario());
+        if (!usuariosService.cpfValidator(usuarios.getCpf())) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new MessageExceptionHandler(new Date(), HttpStatus.NOT_FOUND.value(),"CPF invalido" ));
+        }
         return ResponseEntity.status(HttpStatus.OK).body(usuariosService.save(usuarios));
     }
-
 
     @PostMapping("/login")
     public ResponseEntity<Object> loginUser(@RequestBody LoginDto loginDto) {
